+++ NEWS.md
@@ -1,7 +1,10 @@
 ## tsfgrnn 0.3.0
 
 * Automatic computation of smoothing parameter can be done with rolling origin or fixed origin (less accurate, but faster).
+
 * Lags can be selected using forward selection or backward elimination (feature selection techniques)
+
+* Time series cannot be scaled
 
 ## tsfgrnn 0.2.0